+++ src/CHANGES.md
@@ -6,6 +6,7 @@
 * fix webUI translation #1346
 * fix protection mask #1352
 * merge PR: Add Watchdog for ESP32 #1367
+* merge PR: ETH support for CMT2300A - HMS/HMT #1356
 
 ## 0.8.57 - 2024-01-15
 * merge PR: fix immediate clearing of display after sunset #1364